nixos: Don't format when FS is mounted

+++ bootstrap/build.sh
@@ -3,23 +3,25 @@ set -eo pipefail
 
 source "$(dirname "$0")/setup.sh"
 
-echo "Creating partitions on /dev/${NIX_INSTALL_BLK_DEV}..."
-parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mklabel gpt
-parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart primary ext4 512MiB -8GiB
-parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart primary linux-swap -8GiB 100%
-parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart ESP fat32 1MiB 512MiB
-parted "/dev/${NIX_INSTALL_BLK_DEV}" -- set 3 esp on
+if ! findmnt /mnt > /dev/null; then
+  echo "Creating partitions on /dev/${NIX_INSTALL_BLK_DEV}..."
+  parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mklabel gpt
+  parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart primary ext4 512MiB -8GiB
+  parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart primary linux-swap -8GiB 100%
+  parted "/dev/${NIX_INSTALL_BLK_DEV}" -- mkpart ESP fat32 1MiB 512MiB
+  parted "/dev/${NIX_INSTALL_BLK_DEV}" -- set 3 esp on
 
-echo "Creating filesystems on /dev/${NIX_INSTALL_BLK_DEV}…"
-mkfs.ext4 -L nixos "/dev/"${NIX_INSTALL_BLK_DEV}1
-mkswap -L swap "/dev/${NIX_INSTALL_BLK_DEV}2"
-mkfs.fat -F 32 -n boot "/dev/${NIX_INSTALL_BLK_DEV}3"
+  echo "Creating filesystems on /dev/${NIX_INSTALL_BLK_DEV}…"
+  mkfs.ext4 -L nixos "/dev/"${NIX_INSTALL_BLK_DEV}1
+  mkswap -L swap "/dev/${NIX_INSTALL_BLK_DEV}2"
+  mkfs.fat -F 32 -n boot "/dev/${NIX_INSTALL_BLK_DEV}3"
 
-echo "Mounting partitions under /mnt and enabling swap…"
-mount /dev/disk/by-label/nixos /mnt
-mkdir /mnt/boot
-mount /dev/disk/by-label/boot /mnt/boot
-swapon /dev/disk/by-label/swap
+  echo "Mounting partitions under /mnt and enabling swap…"
+  mount /dev/disk/by-label/nixos /mnt
+  mkdir /mnt/boot
+  mount /dev/disk/by-label/boot /mnt/boot
+  swapon /dev/disk/by-label/swap
+fi
 
 (
     cd "$(dirname "$0")/.."
